requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
transformers==4.50.3
torch==2.6.0
pillow==11.1.0
